protect the user from the user

+++ docker-setup.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
-# make sure *.php are in /var/www/html
+if ! [ -f /.dockerenv ]; then
+    echo "This script is made to be run automatically inside the container."
+    exit
+fi
 cd /var/www
 mv -f *.php ./html/
 chmod a+rw /var/www/html
